Created new function
New function creates an array to hold the different choices in a variable named choices. Number is generated by math.random, multiplying it by 3, and then rounding it down to create a selection between 0, 1, and 2. The choice is returned by using the number stored in the choice variable to be the index of the choices array.