Added logic checks
In the checkWin function, once the user has achieved 5 wins or 5 losses, the global variable win will change from false to true. This prevents the game from being played any further because in the game function, because the if condition win == false no longer passes, the else function is called into just a return.